refactor: 패키지 분리
뷰와 핵심 로직 분리 위해 패키지 분리함.
