+++ src/test/java/racingcar/domain/InputCarCountTest.java
@@ -1,7 +1,8 @@
-package racingcar;
+package racingcar.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.InputCarCount;
 
 import static org.assertj.core.api.Assertions.*;
